% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jm_make_assoc_terms.R
\name{collapse_within_groups}
\alias{collapse_within_groups}
\title{Collapse the linear predictor across the lower level units
clustered an individual, using the function specified in the
'grp_assoc' argument}
\usage{
collapse_within_groups(eta, grp_idx, grp_assoc = "sum")
}
\arguments{
\item{eta}{The linear predictor evaluated for all lower level groups
at the quadrature points.}

\item{grp_idx}{An N*2 array providing the indices of the first (col 1)
and last (col 2) observations in eta that correspond to individuals
i = 1,...,N.}

\item{grp_assoc}{Character string, the function to use to collapse
across the lower level units clustered within individuals.}
}
\value{
A vector or matrix, depending on the method called.
}
\description{
Collapse the linear predictor across the lower level units
clustered an individual, using the function specified in the
'grp_assoc' argument
}
