% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{evaluate_log_survival.default}
\alias{evaluate_log_survival.default}
\title{Evaluate the log baseline hazard at the specified times
given the vector or matrix of MCMC draws for the baseline
hazard coeffients / parameters}
\usage{
\method{evaluate_log_survival}{default}(log_haz, qnodes, qwts)
}
\arguments{
\item{log_haz}{A vector containing the log hazard for each
individual, evaluated at each of the quadrature points. The
vector should be ordered such that the first N elements contain
the log_haz evaluated for each individual at quadrature point 1,
then the next N elements are the log_haz evaluated for each
individual at quadrature point 2, and so on.}

\item{qnodes}{Integer specifying the number of quadrature nodes
at which the log hazard was evaluated for each individual.}

\item{qwts}{A vector of unstandardised GK quadrature weights.}
}
\value{
A vector or matrix of log survival probabilities.
}
\description{
Evaluate the log baseline hazard at the specified times
given the vector or matrix of MCMC draws for the baseline
hazard coeffients / parameters
}
