% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extract_pars}
\alias{extract_pars}
\title{Extract parameters from stanmat and return as a list}
\usage{
extract_pars(object, ...)
}
\arguments{
\item{object}{A stanmvreg or stansurv object}

\item{...}{Additional arguments passed to the method}
}
\value{
A named list
}
\description{
Extract parameters from stanmat and return as a list
}
