% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extract_pars.stansurv}
\alias{extract_pars.stansurv}
\title{Extract parameters from stanmat and return as a list}
\usage{
\method{extract_pars}{stansurv}(object, stanmat = NULL, means = FALSE, ...)
}
\arguments{
\item{object}{A stanmvreg or stansurv object}

\item{stanmat}{A matrix of posterior draws, may be provided if the desired
stanmat is only a subset of the draws from as.matrix(object$stanfit)}

\item{means}{Logical, if TRUE then the posterior means are returned}

\item{...}{Additional arguments passed to the method}
}
\value{
A named list
}
\description{
Extract parameters from stanmat and return as a list
}
