% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmvreg-methods.R
\name{family.stanmvreg}
\alias{family.stanmvreg}
\title{family method for stanmvreg objects}
\usage{
\method{family}{stanmvreg}(object, m = NULL, ...)
}
\arguments{
\item{object, ...}{See \code{\link[stats]{family}}.}

\item{m}{Integer specifying the number or name of the submodel}
}
\description{
family method for stanmvreg objects
}
\keyword{internal}
