% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanreg-methods.R
\name{family.stanreg}
\alias{family.stanreg}
\title{family method for stanreg objects}
\usage{
\method{family}{stanreg}(object, ...)
}
\arguments{
\item{object, ...}{See \code{\link[stats]{family}}.}
}
\description{
family method for stanreg objects
}
\keyword{internal}
