% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanreg-methods.R
\name{formula.stanreg}
\alias{formula.stanreg}
\title{formula method for stanreg objects}
\usage{
\method{formula}{stanreg}(x, ..., m = NULL)
}
\arguments{
\item{x}{A stanreg object.}

\item{...}{Can contain \code{fixed.only} and \code{random.only} arguments 
that both default to \code{FALSE}. Also, for stan_surv models, can contain
\code{remove.tve} which defaults to FALSE, but if TRUE then any 
'tve(varname)' terms in the model formula are returned as 'varname'.}
}
\description{
formula method for stanreg objects
}
\keyword{internal}
