% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_data.R
\name{get_model_data}
\alias{get_model_data}
\title{Return a data frame for each submodel that:
(1) only includes variables used in the model formula
(2) only includes rows contained in the glmod/coxmod model frames
(3) ensures that additional variables that are required
    such as the ID variable or variables used in the
    interaction-type association structures, are included.}
\usage{
get_model_data(object, ...)
}
\arguments{
\item{object}{A stansurv, stanmvreg or stanjm object.}

\item{...}{Additional arguments passed to the method}
}
\value{
A data frame or list of data frames with all the
  (unevaluated) variables required for predictions.
}
\description{
It is necessary to drop unneeded variables though so that
errors are not encountered if the original data contained
NA values for variables unrelated to the model formula.
We generate a data frame here for in-sample predictions
rather than using a model frame, since some quantities will
need to be recalculated at quadrature points etc, for example
in posterior_survfit.
}
