% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_surv}
\alias{get_surv}
\title{Extract survival response from a stansurv or stanjm object}
\usage{
get_surv(object, ...)
}
\arguments{
\item{object}{A \code{stansurv} or \code{stanjm} object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A \code{Surv} object, see \code{?survival::Surv}.
}
\description{
Extract survival response from a stansurv or stanjm object
}
\keyword{internal}
