% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_y}
\alias{get_y}
\alias{get_x}
\alias{get_z}
\title{Extract X, Y or Z from a stanreg object}
\usage{
get_y(object, ...)

get_x(object, ...)

get_z(object, ...)
}
\arguments{
\item{object}{A fitted model object returned by one of the 
\pkg{rstanarm} modeling functions. See \code{\link{stanreg-objects}}.}

\item{...}{Other arguments passed to methods. For a \code{stanmvreg} object
this can be an integer \code{m} specifying the submodel.}
}
\value{
For \code{get_x} and \code{get_z}, a matrix. For \code{get_y}, either
  a vector or a matrix, depending on how the response variable was specified.
}
\description{
Extract X, Y or Z from a stanreg object
}
\keyword{internal}
