% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{linear_predictor.default}
\alias{linear_predictor.default}
\title{Methods for creating linear predictor}
\usage{
\method{linear_predictor}{default}(beta, x, offset = NULL)
}
\arguments{
\item{beta}{A vector or matrix or parameter estimates.}

\item{x}{Predictor matrix.}

\item{offset}{Optional offset vector.}
}
\value{
A vector or matrix.
}
\description{
Make linear predictor vector from x and point estimates for beta, or linear
predictor matrix from x and full posterior sample of beta.
}
