% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{linkinv.character}
\alias{linkinv.character}
\title{Get inverse link function}
\usage{
\method{linkinv}{character}(x, ...)
}
\arguments{
\item{x}{A character string}

\item{...}{Other arguments passed to methods. For a \code{stanmvreg} object
this can be an integer \code{m} specifying the submodel.}
}
\value{
The inverse link function associated with x.
}
\description{
Get inverse link function
}
