% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{linkinv.stanmvreg}
\alias{linkinv.stanmvreg}
\title{Get inverse link function}
\usage{
\method{linkinv}{stanmvreg}(x, m = NULL, ...)
}
\arguments{
\item{x}{A stanmvreg object}

\item{m}{An integer specifying the submodel.}

\item{...}{Other arguments passed to methods. For a \code{stanmvreg} object
this can be an integer \code{m} specifying the submodel.}
}
\value{
The inverse link function associated with x.
}
\description{
Get inverse link function
}
