% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{ll_args.stanjm}
\alias{ll_args.stanjm}
\title{Alternative ll_args method for stanjm objects that allows data and pars to be
passed directly, rather than constructed using pp_data within the ll_args
method. This can be much faster when used in the MH algorithm within
posterior_survfit, since it doesn't require repeated calls to pp_data.}
\usage{
\method{ll_args}{stanjm}(object, data, pars, m = 1, reloo_or_kfold = FALSE, ...)
}
\arguments{
\item{object}{A stanmvreg object}

\item{data}{Output from .pp_data_jm}

\item{pars}{Output from extract_pars}

\item{m}{Integer specifying which submodel}

\item{reloo_or_kfold}{logical. TRUE if ll_args is for reloo or kfold}

\item{...}{Additional arguments passed to the method}
}
\description{
Alternative ll_args method for stanjm objects that allows data and pars to be
passed directly, rather than constructed using pp_data within the ll_args
method. This can be much faster when used in the MH algorithm within
posterior_survfit, since it doesn't require repeated calls to pp_data.
}
