% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{ll_args.stansurv}
\alias{ll_args.stansurv}
\title{get arguments needed for ll_fun}
\usage{
\method{ll_args}{stansurv}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{stansurv object}

\item{newdata}{same as posterior predict}

\item{...}{For models without group-specific terms (i.e., not stan_[g]lmer),
if reloo_or_kfold is TRUE and 'newdata' is specified then ... is used to
pass 'newx' and 'stanmat' from reloo or kfold (bypassing pp_data). This is a
workaround in case there are issues with newdata containing factors with
only a single level. Or for stanmvreg objects, then ... can be used to pass
'stanmat', which may be a matrix with a reduced number of draws (potentially
just a single MCMC draw).}
}
\value{
a named list with elements data, draws, S (posterior sample size) and
  N = number of observations
}
\description{
get arguments needed for ll_fun
}
