% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{logit}
\alias{logit}
\alias{invlogit}
\title{Logit and inverse logit}
\usage{
logit(x)

invlogit(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
A numeric vector the same length as \code{x}.
}
\description{
Logit and inverse logit
}
