% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmvreg-methods.R
\name{model.frame.stanmvreg}
\alias{model.frame.stanmvreg}
\title{model.frame method for stanmvreg objects}
\usage{
\method{model.frame}{stanmvreg}(formula, fixed.only = FALSE, m = NULL, ...)
}
\arguments{
\item{formula, ...}{See \code{\link[stats]{model.frame}}.}

\item{fixed.only}{See \code{\link[lme4:merMod-class]{model.frame.merMod}}.}

\item{m}{Integer specifying the number or name of the submodel}
}
\description{
model.frame method for stanmvreg objects
}
\keyword{internal}
