% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanreg-methods.R
\name{model.frame.stanreg}
\alias{model.frame.stanreg}
\title{model.frame method for stanreg objects}
\usage{
\method{model.frame}{stanreg}(formula, fixed.only = FALSE, ...)
}
\arguments{
\item{formula, ...}{See \code{\link[stats]{model.frame}}.}

\item{fixed.only}{See \code{\link[lme4:merMod-class]{model.frame.merMod}}.}
}
\description{
model.frame method for stanreg objects
}
\keyword{internal}
