% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanreg-methods.R
\name{model.matrix.stanreg}
\alias{model.matrix.stanreg}
\title{model.matrix method for stanreg objects}
\usage{
\method{model.matrix}{stanreg}(object, ...)
}
\arguments{
\item{object, ...}{See \code{\link[stats]{model.matrix}}.}
}
\description{
model.matrix method for stanreg objects
}
\keyword{internal}
