% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_linpred.R
\name{posterior_linpred}
\alias{posterior_linpred}
\title{Posterior distribution of the linear predictor}
\usage{
posterior_linpred(object, transform = FALSE, newdata = NULL,
  re.form = NULL, offset = NULL, XZ = FALSE, ...)
}
\arguments{
\item{object}{A fitted model object returned by one of the
\pkg{rstanarm} modeling functions. See \code{\link{stanreg-objects}}.}

\item{transform}{Should the linear predictor be transformed using the
inverse-link function? The default is \code{FALSE}, in which case the 
untransformed linear predictor is returned.}

\item{newdata, re.form, offset}{Same as for \code{\link{posterior_predict}}.}

\item{XZ}{If \code{TRUE} then instead of computing the linear predictor the 
design matrix \code{X} (or \code{cbind(X,Z)} for models with group-specific
terms) constructed from \code{newdata} is returned. The default is 
\code{FALSE}.}

\item{...}{Currently unused.}
}
\value{
The default is to return a \code{draws} by \code{nrow(newdata)} 
  matrix of simulations from the posterior distribution of the (possibly 
  transformed) linear predictor. The exception is if the argument \code{XZ} 
  is set to \code{TRUE} (see the \code{XZ} argument description above).
}
\description{
Extract the posterior draws of the linear predictor, possibly transformed by 
the inverse-link function. This function is occasionally convenient, but it 
should be used sparingly. Inference and model checking should generally be 
carried out using the posterior predictive distribution (see
\code{\link{posterior_predict}}).
}
\examples{
if (!exists("example_model")) example(example_model)

# linear predictor on log-odds scale
linpred <- posterior_linpred(example_model)
# probabilities
probs <- posterior_linpred(example_model, transform = TRUE)

# not conditioning on any group-level parameters
probs2 <- posterior_linpred(example_model, transform = TRUE, re.form = NA)

}
\seealso{
\code{\link{posterior_predict}} to draw from the posterior 
  predictive distribution of the outcome, which is almost always preferable.
}
\keyword{internal}

