% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check.R
\name{ppcheck}
\alias{ppcheck}
\title{ppcheck (deprecated)}
\usage{
ppcheck(object, check = "distributions", nreps = NULL, overlay = TRUE,
  test = "mean", ...)
}
\arguments{
\item{object}{A fitted model object returned by one of the
\pkg{rstanarm} modeling functions. See \code{\link{stanreg-objects}}.}

\item{check}{The type of plot (possibly abbreviated) to show. One of 
\code{"distributions"}, \code{"residuals"}, \code{"scatter"}, 
\code{"test"}. See Details for descriptions.}

\item{nreps}{The number of \eqn{y^{rep}}{yrep} datasets to generate from the
posterior predictive distribution (\code{\link{posterior_predict}}) and
show in the plots. The default is \code{nreps=3} for
\code{check="residuals"} and \code{nreps=8} for
\code{check="distributions"}. If \code{check="test"}, \code{nreps} is
ignored and the number of simulated datasets is the number of post-warmup
draws from the posterior distribution. If \code{check="scatter"},
\code{nreps} is not ignored but defaults to the number of post-warmup
draws.}

\item{overlay}{For \code{check="distributions"} only, should distributions be
plotted as density estimates overlaid in a single plot (\code{TRUE}, the 
default) or as separate histograms (\code{FALSE})?}

\item{test}{For \code{check="test"} only, a character vector (of length 1 or 
2) naming a single function or a pair of functions. The function(s) should 
take a vector input and return a scalar test statistic. See Details.}

\item{...}{Optional arguments to geoms to control features of the plots 
(e.g. \code{binwidth} if the plot is a histogram).}
}
\description{
ppcheck is deprecated and will be removed in the near future. Use
\code{\link{pp_check}} instead.
}
\keyword{internal}

