% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.stanreg}
\alias{predict.stanreg}
\title{Predict method for stanreg objects}
\usage{
\method{predict}{stanreg}(
  object,
  ...,
  newdata = NULL,
  type = c("link", "response"),
  se.fit = FALSE
)
}
\arguments{
\item{object}{A fitted model object returned by one of the 
\pkg{rstanarm} modeling functions. See \code{\link{stanreg-objects}}.}

\item{...}{Ignored.}

\item{newdata}{Optionally, a data frame in which to look for variables with 
which to predict. If omitted, the model matrix is used.}

\item{type}{The type of prediction. The default \code{'link'} is on the scale
of the linear predictors; the alternative \code{'response'} is on the scale
of the response variable.}

\item{se.fit}{A logical scalar indicating if standard errors should be 
returned. The default is \code{FALSE}.}
}
\value{
A vector if \code{se.fit} is \code{FALSE} and a list if \code{se.fit}
  is \code{TRUE}.
}
\description{
This method is primarily intended to be used only for models fit using 
optimization. For models fit using MCMC or one of the variational
approximations, see \code{\link{posterior_predict}}.
}
\seealso{
\code{\link{posterior_predict}}
}
