% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{print.kfold}
\alias{print.kfold}
\title{Print method for kfold}
\usage{
\method{print}{kfold}(x, digits = 1, ...)
}
\arguments{
\item{x, digits, ...}{See \code{\link{print}}.}
}
\description{
Print method for kfold
}
\keyword{internal}

