% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_survfit.R
\name{print.survfit.stanjm}
\alias{print.survfit.stanjm}
\title{Generic print method for \code{survfit.stanjm} objects}
\usage{
\method{print}{survfit.stanjm}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{survfit.stanjm}, returned by a call to 
\code{\link{posterior_survfit}}.}

\item{digits}{Number of digits to use for formatting the time variable and 
the survival probabilities.}

\item{...}{Ignored.}
}
\description{
Generic print method for \code{survfit.stanjm} objects
}
\keyword{internal}
