% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/priors.R
\name{priors}
\alias{cauchy}
\alias{normal}
\alias{priors}
\alias{student_t}
\title{Prior distributions}
\usage{
normal(location = 0, scale = NULL)

student_t(df = 1, location = 0, scale = NULL)

cauchy(location = 0, scale = NULL)
}
\arguments{
\item{location}{prior location. Defaults to 0. For \code{normal} and
\code{student_t} (provided that \code{df > 1}) this is the prior mean. For
\code{cauchy} the mean is undefined and \code{location} is the prior median.}

\item{scale}{prior scale. Default depends (see Details).}

\item{df}{prior degrees of freedom. Defaults to 1, in which case
\code{student_t} is equivalent to \code{cauchy}.}
}
\description{
These functions are used to specify the \code{prior} and
\code{prior.for.intercept} arguments of the \code{stan_regression},
\code{stan_lm}, and \code{stan_glm} functions.
}
\details{
If \code{scale} is not specified it will default to 10 for the
intercept and 2.5 for the other coefficients, unless the probit link function
is used, in which case these defaults are scaled by a factor of
\code{dnorm(0)/dlogis(0)} (roughly 1.6).
}
\examples{
\dontrun{
stan_lm(y ~ x1 + x2, prior = student_t(4, 0, 2.5), prior.for.intercept = cauchy(0,10))
}
}

