% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo-prediction.R
\name{psis.stanreg}
\alias{psis.stanreg}
\title{Perform importance sampling with stanreg objects}
\usage{
\method{psis}{stanreg}(log_ratios, ...)
}
\arguments{
\item{log_ratios}{log_ratios for PSIS computation}

\item{...}{Additional arguments passed to \code{\link[loo]{psis}}}
}
\description{
Perform importance sampling with stanreg objects
}
