% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{quadrature_sum.default}
\alias{quadrature_sum.default}
\title{Apply quadrature weights and sum over nodes}
\usage{
\method{quadrature_sum}{default}(x, qnodes, qwts)
}
\arguments{
\item{x}{Inputs}

\item{qnodes}{Nodes}

\item{qwts}{Quadrature weights}
}
\description{
Apply quadrature weights and sum over nodes
}
