% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanreg_list.R
\name{rename_loos}
\alias{rename_loos}
\title{loo/waic/kfold objects created by rstanarm have a model_name attribute.
when a stanreg_list is created those attributes should be changed to match
the names of the models used for the stanreg_list in case user has specified
the model_names argument}
\usage{
rename_loos(x, ...)
}
\arguments{
\item{x}{loo/waic/kfold object}

\item{...}{Additional arguments}
}
\value{
Renamed object
}
\description{
loo/waic/kfold objects created by rstanarm have a model_name attribute.
when a stanreg_list is created those attributes should be changed to match
the names of the models used for the stanreg_list in case user has specified
the model_names argument
}
