% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanreg_list.R
\name{rename_loos.stanreg}
\alias{rename_loos.stanreg}
\title{Change model_name attributes of a loo/waic/kfold object stored in a stanreg object,}
\usage{
\method{rename_loos}{stanreg}(x, new_model_name, ...)
}
\arguments{
\item{x}{loo/waic/kfold object}

\item{new_model_name}{new name for object}

\item{...}{Additional arguments}
}
\value{
Renamed object
}
\description{
Change model_name attributes of a loo/waic/kfold object stored in a stanreg object,
}
