% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanreg_list.R
\name{rename_loos.stanreg_list}
\alias{rename_loos.stanreg_list}
\title{Change model_name attributes of loo/waic/kfold objects to correspond to
model names used for stanreg_list}
\usage{
\method{rename_loos}{stanreg_list}(x, ...)
}
\arguments{
\item{x}{loo/waic/kfold object}

\item{...}{Additional arguments}
}
\value{
Renamed object
}
\description{
Change model_name attributes of loo/waic/kfold objects to correspond to
model names used for stanreg_list
}
