% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-rstanarm-deprecated.R
\name{rstanarm-deprecated}
\alias{rstanarm-deprecated}
\alias{prior_options}
\title{Deprecated functions}
\usage{
prior_options(
  prior_scale_for_dispersion = 5,
  min_prior_scale = 1e-12,
  scaled = TRUE
)
}
\arguments{
\item{prior_scale_for_dispersion, min_prior_scale, scaled}{Arguments to 
deprecated \code{prior_options} function. The functionality provided 
by the now deprecated \code{prior_options} function has been replaced 
as follows: 
\describe{
\item{\code{prior_scale_for_dispersion}}{
 Instead of using the \code{prior_scale_for_dispersion} argument to 
 \code{prior_options}, priors for these parameters can now be 
 specified directly when calling \code{\link{stan_glm}} (or
 \code{\link{stan_glmer}}, etc.) using the new \code{prior_aux}
 argument.
}
\item{\code{scaled}}{
 Instead of setting \code{prior_options(scaled=FALSE)}, internal rescaling
 is now toggled using the new \code{autoscale} arguments to
 \code{\link{normal}}, \code{\link{student_t}}, and \code{\link{cauchy}} 
 (the other prior distributions do not support 'autoscale').
}
\item{\code{min_prior_scale}}{
 No replacement. \code{min_prior_scale} (the minimum possible scale
 parameter value that be used for priors) is now fixed to \code{1e-12}.
}
}}
}
\description{
These functions are deprecated and will be removed in a future release. The 
\strong{Arguments} section below provides details on how the functionality 
obtained via each of the arguments has been replaced.
}
