% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinystan.R
\name{shinystan}
\alias{launch_shinystan}
\alias{shinystan}
\title{Using the ShinyStan GUI with stanreg objects}
\description{
The \code{\link[shinystan]{launch_shinystan}} function will accept a 
\link[=stanreg-objects]{'stanreg' object} as input. Currently, almost any
model fit using one of \pkg{rstanarm}'s model-fitting functions can be
used with ShinyStan. The only exception is that ShinyStan does not currently
support \pkg{rstanarm} models fit using \code{algorithm='optimizing'}.
}
\details{
See the \pkg{\link[=shinystan-package]{shinystan}} package documentation for
more information.
}
\examples{
\dontrun{
# Launch the ShinyStan app (saving resulting shinystan object as sso)
sso <- launch_shinystan(example_model)
}


}

