% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanreg-methods.R
\name{sigma}
\alias{sigma}
\title{Extract residual standard deviation}
\usage{
sigma(object, ...)
}
\arguments{
\item{object}{Fitted model object.}

\item{...}{Arguments to methods.}
}
\description{
Extract residual standard deviation
}
\keyword{internal}

