% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{split2.matrix}
\alias{split2.matrix}
\title{Split a vector or matrix into a specified number of segments and return
each segment as an element of a list. The matrix method allows splitting
across the column (bycol = TRUE) or row margin (bycol = FALSE).}
\usage{
\method{split2}{matrix}(x, n_segments = 1, bycol = TRUE, ...)
}
\arguments{
\item{x}{A matrix}

\item{n_segments}{Integer specifying the number of segments.}

\item{bycol}{Logical, should a matrix be split along the column or row margin?}

\item{...}{Additional arguments passed to the method.}
}
\description{
Split a vector or matrix into a specified number of segments and return
each segment as an element of a list. The matrix method allows splitting
across the column (bycol = TRUE) or row margin (bycol = FALSE).
}
