% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanreg_list.R
\name{stanreg_list}
\alias{stanreg_list}
\alias{stanmvreg_list}
\alias{stanjm_list}
\title{Create a list of fitted model objects}
\usage{
stanreg_list(...)

stanmvreg_list(...)

stanjm_list(...)
}
\arguments{
\item{...}{One or more fitted model objects.}
}
\value{
A list of class \code{"stanreg_list"}, \code{"stanmvreg_list"}, or
  \code{"stanjm_list"}, containing the fitted model objects and some metadata
  stored as attributes.
}
\description{
Create a list of fitted model objects
}
\seealso{
\code{\link{loo_model_weights}} for usage of \code{stanreg_list}.
}
