% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{stanreg_methods}
\alias{confint.stanreg}
\alias{log_lik.stanreg}
\alias{print.stanreg}
\alias{residuals.stanreg}
\alias{se.stanreg}
\alias{stanreg_methods}
\alias{summary.stanreg}
\alias{vcov.stanreg}
\title{Methods}
\usage{
\method{residuals}{stanreg}(object, ...)

\method{vcov}{stanreg}(object, ...)

\method{se}{stanreg}(object, parm)

\method{confint}{stanreg}(object, parm, level = 0.95, ...)

\method{print}{stanreg}(x, ...)

\method{summary}{stanreg}(object, ...)

\method{log_lik}{stanreg}(object)
}
\arguments{
\item{object,x}{a model fit with \code{\link{stan_lm}} or
\code{\link{stan_glm}}.}

\item{...}{other arguments to \code{print} or \code{summary}. See Details.}

\item{parm}{a character vector of parameter names.}

\item{level}{confidence level.}
}
\description{
Methods
}
\note{
Unlike \code{\link[stats]{glm}}, residuals are of type \code{'response'}
not \code{'deviance'} (see \code{\link[stats]{residuals.glm}}).
}

