% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmvreg-methods.R
\name{terms.stanmvreg}
\alias{terms.stanmvreg}
\title{terms method for stanmvreg objects}
\usage{
\method{terms}{stanmvreg}(x, fixed.only = TRUE, random.only = FALSE, m = NULL, ...)
}
\arguments{
\item{x, fixed.only, random.only, ...}{See lme4:::terms.merMod.}

\item{m}{Integer specifying the number or name of the submodel}
}
\description{
terms method for stanmvreg objects
}
\keyword{internal}
