% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanreg-methods.R
\name{terms.stanreg}
\alias{terms.stanreg}
\title{terms method for stanreg objects}
\usage{
\method{terms}{stanreg}(x, ..., fixed.only = TRUE, random.only = FALSE)
}
\arguments{
\item{x, fixed.only, random.only, ...}{See lme4:::terms.merMod.}
}
\description{
terms method for stanreg objects
}
\keyword{internal}
