% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{truncate.numeric}
\alias{truncate.numeric}
\title{Method to truncate a numeric vector at defined limits}
\usage{
\method{truncate}{numeric}(con, lower = NULL, upper = NULL, ...)
}
\arguments{
\item{con}{A numeric vector.}

\item{lower}{Scalar, the lower limit for the returned vector.}

\item{upper}{Scalar, the upper limit for the returned vector.}

\item{...}{Additional arguments passed to the method.}
}
\value{
A numeric vector.
}
\description{
Method to truncate a numeric vector at defined limits
}
