% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_surv.R
\name{tve}
\alias{tve}
\title{Time-varying effects in Bayesian survival models}
\usage{
tve(x, type = "bs", df = NULL, knots = NULL, degree = 3L)
}
\arguments{
\item{x}{The covariate for which a time-varying coefficient should be 
estimated.}

\item{type}{The type of function used to model the time-varying coefficient.
Currently only \code{type = "bs"} is allowed. This corresponds to a
B-spline function. Note that \emph{cubic} B-splines are used by default
but this can be changed by the user via the \code{degree} argument 
described below. Of particular note is that \code{degree = 0} is 
is treated as a special case corresponding to a piecewise constant basis.}

\item{df}{A positive integer specifying the degrees of freedom 
for the B-spline function. Two boundary knots and \code{df - degree} 
internal knots are used to generate the B-spline function.
The internal knots are placed at equally spaced percentiles of the 
distribution of the uncensored event times. The default is to use 
\code{df = 3} unless \code{df} or \code{knots} is explicitly 
specified by the user.}

\item{knots}{A numeric vector explicitly specifying internal knot
locations for the B-spline function. Note that \code{knots} cannot be 
specified if \code{df} is specified. Also note that this argument only 
controls the \emph{internal} knot locations. In addition, boundary 
knots are placed at the earliest entry time and latest event or 
censoring time and these cannot be changed by the user.}

\item{degree}{A positive integer specifying the degree for the B-spline 
function. The order of the B-spline is equal to \code{degree + 1}.
Note that \code{degree = 0} is allowed and is treated as a special
case corresponding to a piecewise constant basis.}
}
\value{
The exported \code{tve} function documented here just returns 
  \code{x}. However, when called internally the \code{tve} function returns 
  several other pieces of useful information. For the most part, these are 
  added to the formula element of the returned \code{\link{stanreg-objects}}
  (that is \code{object[["formula"]]} where \code{object} is the fitted
  model). Information added to the formula element of the \code{stanreg} 
  object includes the following:
  \itemize{
  \item \code{tt_vars}: A list with the names of variables in the model
  formula that were wrapped in the \code{tve} function.
  \item \code{tt_types}: A list with the \code{type} (e.g. \code{"bs"}) 
  of \code{tve} function corresponding to each variable in \code{tt_vars}.
  \item \code{tt_degrees}: A list with the \code{degree} for the
  B-spline function corresponding to each variable in \code{tt_vars}.
  \item \code{tt_calls}: A list with the call required to construct the
  transformation of time for each variable in \code{tt_vars}.
  \item \code{tt_forms}: Same as \code{tt_calls} but expressed as formulas.
  \item \code{tt_frame}: A single formula that can be used to generate a 
  model frame that contains the unique set of transformations of time 
  (i.e. the basis terms) that are required to build all time-varying 
  coefficients in the model. In other words a single formula with the 
  unique element(s) contained in \code{tt_forms}.
  }
}
\description{
This is a special function that can be used in the formula of a Bayesian 
survival model estimated using \code{\link{stan_surv}}. It specifies that a 
time-varying coefficient should be estimated for the covariate \code{x}. 
The time-varying coefficient is currently modelled using B-splines (with
piecewise constant included as a special case). Note that the \code{tve} 
function only has meaning when evaluated within the formula of a 
\code{\link{stan_surv}} call and does not have meaning outside of that 
context. The exported function documented here just returns \code{x}. 
However when called internally the \code{tve} function returns several 
other pieces of useful information used in the model fitting.
}
\examples{
# Exported function just returns the input variable
identical(pbcSurv$trt, tve(pbcSurv$trt)) # returns TRUE

# Internally the function returns and stores information 
# used to form the time-varying coefficients in the model
m1 <- stan_surv(Surv(futimeYears, death) ~ tve(trt) + tve(sex, degree = 0),
                data = pbcSurv, chains = 1, iter = 50)
m1$formula[["tt_vars"]]
m1$formula[["tt_forms"]]
 
}
