% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_glm.fit.R
\name{unpad_reTrms.array}
\alias{unpad_reTrms.array}
\title{Drop the extra reTrms from an array x}
\usage{
\method{unpad_reTrms}{array}(x, columns = TRUE, ...)
}
\arguments{
\item{x}{An array (e.g. the posterior sample or matrix of summary
stats)}

\item{columns}{Do the columns (TRUE) or rows (FALSE) correspond to the
variables?}

\item{...}{Additional arguments (not used)}
}
\description{
Drop the extra reTrms from an array x
}
