% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_glm.fit.R
\name{unpad_reTrms.default}
\alias{unpad_reTrms.default}
\title{Drop the extra reTrms from a matrix x}
\usage{
\method{unpad_reTrms}{default}(x, ...)
}
\arguments{
\item{x}{A matrix (e.g. the posterior sample or matrix of summary
stats)}

\item{...}{Additional arguments (not used)}
}
\description{
Drop the extra reTrms from a matrix x
}
