% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_glm.fit.R
\name{unpad_reTrms}
\alias{unpad_reTrms}
\title{Drop the extra reTrms from a matrix x}
\usage{
unpad_reTrms(x, ...)
}
\arguments{
\item{x}{A matrix or array (e.g. the posterior sample or matrix of summary
stats)}

\item{...}{Additional arguments (not used)}
}
\description{
Drop the extra reTrms from a matrix x
}
