% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanreg-methods.R
\name{formula.stanreg}
\alias{formula.stanreg}
\title{formula method for stanreg objects}
\usage{
\method{formula}{stanreg}(x, ..., m = NULL)
}
\arguments{
\item{x}{A stanreg object.}

\item{...}{Can contain \code{fixed.only} and \code{random.only} arguments 
that both default to \code{FALSE}.}
}
\description{
formula method for stanreg objects
}
\keyword{internal}
