% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{log_lik}
\alias{log_lik}
\title{Pointwise log-likelihood}
\usage{
log_lik(object)
}
\arguments{
\item{object}{object to use.}
}
\value{
pointwise log-likelihood matrix
}
\description{
Pointwise log-likelihood
}
\keyword{internal}

