% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/log_lik.R
\name{log_lik}
\alias{log_lik}
\title{Pointwise log-likelihood}
\usage{
log_lik(object)
}
\arguments{
\item{object}{a model fit with \code{\link{stan_lm}} or
\code{\link{stan_glm}}.}
}
\value{
a matrix.
}
\description{
Compute the \eqn{S} by \eqn{N} pointwise log-likelihood matrix, where \eqn{S}
is the size of the posterior sample (the number of simulations) and \eqn{N}
is the number of data points.
}

