% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pairs.stanreg}
\alias{pairs.stanreg}
\title{Pairs method for stanreg objects}
\usage{
\method{pairs}{stanreg}(x, ...)
}
\arguments{
\item{x}{A stanreg object returned by one of the rstanarm modeling functions.}

\item{...}{Arguments to pass to \code{\link[rstan]{pairs.stanfit}}.}
}
\description{
See \code{\link[rstan]{pairs.stanfit}} for details.
}
\details{
See the Details section in \code{\link[rstan]{pairs.stanfit}}.
}
\examples{
if (!exists("example_model")) example(example_model)
pairs(example_model, pars = c("(Intercept)", "log-posterior"))

}

