% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jm_data_block.R
\docType{import}
\name{reexports}
\alias{Surv}
\alias{reexports}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{survival}{\code{\link[survival]{Surv}}}
}}
\keyword{internal}

