\name{internal-rstanarm}
\title{Internal objects of the package}
\description{
There are not any internal objects to worry about but this entry is used to 
silently trigger the build for the other documentation pages.
\Sexpr[results=hide,stage=build]{roxygen2::roxygenize(clean = TRUE)}
}
\keyword{internal}
