% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{se}
\alias{se}
\title{Standard errors}
\usage{
se(object, parm)
}
\arguments{
\item{object}{object to use.}

\item{parm}{A character vector of parameter names.}
}
\value{
estimated standard errors for the parameters named in \code{parm}.
}
\description{
Standard errors
}
\keyword{internal}

