% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanreg-methods.R
\name{se}
\alias{se}
\title{Extract standard errors}
\usage{
se(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Arguments to methods.}
}
\value{
Standard errors of model parameters.
}
\description{
Generic function for extracting standard errors from fitted models.
}
\seealso{
\code{\link{se.stanreg}}
}
\keyword{internal}
