% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_gamm4.R
\name{stan_gamm4}
\alias{plot_nonlinear}
\alias{stan_gamm4}
\title{Bayesian generalized linear additive models with group-specific terms via
Stan}
\usage{
stan_gamm4(formula, random = NULL, family = gaussian(), data = list(),
  weights = NULL, subset = NULL, na.action, knots = NULL,
  drop.unused.levels = TRUE, ..., prior = normal(),
  prior_intercept = normal(), prior_ops = prior_options(),
  prior_covariance = decov(), prior_PD = FALSE, algorithm = c("sampling",
  "meanfield", "fullrank"), adapt_delta = NULL, QR = FALSE,
  sparse = FALSE)

plot_nonlinear(x, smooths, prob = 0.9, facet_args = list(), ...,
  alpha = 1, size = 0.75)
}
\arguments{
\item{formula, random, family, data, knots, drop.unused.levels}{Same as for 
\code{\link[gamm4]{gamm4}}.}

\item{subset, weights, na.action}{Same as \code{\link[stats]{glm}}, 
but rarely specified.}

\item{...}{Further arguments passed to \code{\link[rstan]{sampling}} (e.g. 
\code{iter}, \code{chains}, \code{cores}, etc.) or to
\code{\link[rstan]{vb}} (if \code{algorithm} is \code{"meanfield"} or
\code{"fullrank"}).}

\item{prior}{The prior distribution for the regression coefficients. 
\code{prior} can be a call to \code{normal}, \code{student_t}, 
\code{cauchy}, \code{hs} or \code{hs_plus}. See \code{\link{priors}} for 
details. To omit a prior ---i.e., to use a flat (improper) uniform prior---
\code{prior} can be set to \code{NULL}, although this is rarely a good 
idea. (\strong{Note:} unless \code{QR=TRUE}, if the \code{scaled} argument
to \code{\link{prior_options}} is left at its default and recommended value
of \code{TRUE}, then the scale(s) of \code{prior} may be modified
internally based on the scales of the predictors, as in the \pkg{arm}
package. See \code{\link{priors}} for details on the rescaling and 
\code{\link{prior_summary}} for a summary of the priors used for a
particular model.)}

\item{prior_intercept}{The prior distribution for the intercept. 
\code{prior_intercept} can be a call to \code{normal}, \code{student_t} or 
\code{cauchy}. See \code{\link{priors}} for details. To to omit a prior 
---i.e., to use a flat (improper) uniform prior--- set 
\code{prior_intercept} to \code{NULL}. (\strong{Note:} if a dense 
representation of the design matrix is utilized ---i.e., if the
\code{sparse} argument is left at its default value of \code{FALSE}--- then
the prior distribution for the intercept is set so it applies to the value
when all predictors are centered.)}

\item{prior_ops}{Additional options related to prior distributions. Can 
be \code{NULL} to omit a prior on the dispersion and see 
\code{\link{prior_options}} otherwise.}

\item{prior_covariance}{Cannot be \code{NULL}; see \code{\link{decov}} for
more information about the default arguments.}

\item{prior_PD}{A logical scalar (defaulting to \code{FALSE}) indicating
whether to draw from the prior predictive distribution instead of
conditioning on the outcome.}

\item{algorithm}{A string (possibly abbreviated) indicating the 
estimation approach to use. Can be \code{"sampling"} for MCMC (the
default), \code{"optimizing"} for optimization, \code{"meanfield"} for
variational inference with independent normal distributions, or
\code{"fullrank"} for variational inference with a multivariate normal
distribution. See \code{\link{rstanarm-package}} for more details on the
estimation algorithms. NOTE: not all fitting functions support all four
algorithms.}

\item{adapt_delta}{Only relevant if \code{algorithm="sampling"}. See 
\code{\link{adapt_delta}} for details.}

\item{QR}{A logical scalar (defaulting to \code{FALSE}) but if \code{TRUE}
applies a scaled \code{\link{qr}} decomposition to the design matrix, 
\eqn{X = Q^\ast R^\ast}{X = Q* R*}, where 
\eqn{Q^\ast = Q \sqrt{n-1}}{Q* = Q (n-1)^0.5} and
\eqn{R^\ast = \frac{1}{\sqrt{n-1}} R}{R* = (n-1)^(-0.5) R}. The coefficients
relative to \eqn{Q^\ast}{Q*} are obtained and then premultiplied by the
inverse of \eqn{R^{\ast}}{R*} to obtain coefficients relative to the
original predictors, \eqn{X}. These transformations do not change the 
likelihood of the data but are recommended for computational reasons when 
there are multiple predictors. However, because the coefficients relative
to \eqn{Q^\ast}{Q*} are not very interpretable it is hard to specify an 
informative prior. Setting \code{QR=TRUE} is therefore only recommended 
if you do not have an informative prior for the regression coefficients.}

\item{sparse}{A logical scalar (defaulting to \code{FALSE}) indicating
whether to use a sparse representation of the design (X) matrix. 
Setting this to \code{TRUE} will likely be twice as slow, even if the
design matrix has a considerable number of zeros, but it may allow the
model to be estimated when the computer has too little RAM to
utilize a dense design matrix. If \code{TRUE}, the the design matrix
is not centered (since that would destroy the sparsity) and it is
not possible to specify both \code{QR = TRUE} and \code{sparse = TRUE}.}

\item{x}{An object produced by \code{stan_gamm4}.}

\item{smooths}{An optional character vector specifying a subset of the smooth
functions specified in the call to \code{stan_gamm4}. The default is
include all smooth terms.}

\item{prob}{A scalar between 0 and 1 governing the width of the uncertainty
interval.}

\item{facet_args}{An optional named list of arguments passed to 
\code{\link[ggplot2]{facet_wrap}} (other than the \code{facets} argument).}

\item{alpha, size}{Passed to \code{\link[ggplot2]{geom_ribbon}}.}
}
\value{
A \link[=stanreg-objects]{stanreg} object is returned 
for \code{stan_gamm4}.

\code{plot_nonlinear} returns a ggplot object.
}
\description{
Bayesian inference for GAMMs with flexible priors.
}
\details{
The \code{stan_gamm4} function is similar in syntax to 
  \code{\link[gamm4]{gamm4}} in the \pkg{gamm4} package, which accepts a syntax 
  that is similar to (but not quite as extensive as) that for 
  \code{\link[mgcv]{gamm}} in the \pkg{mgcv} package and converts 
  it internally into the syntax accepted by \code{\link[lme4]{glmer}} in the
  \pkg{lme4} package. But rather than performing (restricted) maximum likelihood 
  estimation, the \code{stan_gamm4} function utilizes MCMC to perform Bayesian 
  estimation. The Bayesian model adds independent priors on the common regression 
  coefficients (in the same way as \code{\link{stan_glm}}) and priors on the 
  terms of a decomposition of the covariance matrices of the group-specific 
  parameters, including the smooths. Estimating these models via MCMC avoids
  the optimization issues that often crop up with GAMMs and provides better
  estimates for the uncertainty in the parameter estimates. 
  
  See \code{\link[gamm4]{gamm4}} for more information about the model
  specicification and \code{\link{priors}} for more information about the
  priors.
  
  The \code{plot_nonlinear} function creates a ggplot object with one facet for
  each smooth function specified in the call to \code{stan_gamm4}. A subset of the
  smooth functions can be specified using the \code{smooths} argument. The 
  plot is conceptually similar to \code{\link[mgcv]{plot.gam}} except the 
  outer lines here demark the edges of posterior uncertainty intervals 
  (credible intervals) rather than confidence intervals and the inner line
  is the posterior median of the function rather than the function implied
  by a point estimate. To change the colors used in the plot see 
  \code{\link[bayesplot]{color_scheme_set}}.
}
\examples{
# from example(gamm4, package = "gamm4"), prefixing gamm4() call with stan_
\donttest{
dat <- mgcv::gamSim(1, n = 400, scale = 2) ## simulate 4 term additive truth
## Now add 20 level random effect `fac'...
dat$fac <- fac <- as.factor(sample(1:20, 400, replace = TRUE))
dat$y <- dat$y + model.matrix(~ fac - 1) \%*\% rnorm(20) * .5

br <- stan_gamm4(y ~ s(x0) + x1 + s(x2), data = dat, random = ~ (1 | fac), 
                 QR = TRUE, chains = 1)
print(br)
plot_nonlinear(br)
plot_nonlinear(br, smooths = "s(x0)", alpha = 2/3)
}
}
\references{
Crainiceanu, C., Ruppert D., and Wand, M. (2005). Bayesian Analysis for 
Penalized Spline Regression Using WinBUGS. 
\emph{Journal of Statistical Software}. \strong{14}(14), 1--22.
\url{https://www.jstatsoft.org/article/view/v014i14}
}
\seealso{
\code{\link{stanreg-methods}} and 
\code{\link[gamm4]{gamm4}}.
}

