% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan_glm.R, R/stan_lm.R
\name{stan_glm}
\alias{stan_glm}
\alias{stan_lm}
\title{Fitting Bayesian generalized linear models via Stan}
\usage{
stan_glm(formula, family = gaussian(), data, weights, subset,
  na.action = NULL, start = NULL, offset = NULL, model = TRUE,
  x = FALSE, y = TRUE, contrasts = NULL, prior = normal(),
  prior.for.intercept = normal(), prior.options = prior_options(), ...)

stan_lm(formula, data, weights, subset, na.action = NULL, start = NULL,
  offset = NULL, model = TRUE, x = FALSE, y = TRUE, contrasts = NULL,
  prior = normal(), prior.for.intercept = normal(),
  prior.options = prior_options(), ...)
}
\arguments{
\item{formula,family,data,subset,x,y,model}{Same as
\code{\link[stats]{glm}}.}

\item{na.action,weights,offset,contrasts}{Same as
\code{\link[stats]{glm}}.}

\item{start}{Same as \code{\link[stats]{glm}}, but if not \code{NULL} also
  used as starting values for the MCMC. If \code{NULL} (the default), then
  \code{\link[rstan]{stan}} is initialized with \code{init = 'random'}.}

\item{prior}{Prior for coefficients. See \code{\link{priors}}.}

\item{prior.for.intercept}{Prior for intercept. See \code{\link{priors}}.}

\item{prior.options}{Additional options related to prior distributions. See
  \code{\link{priors}}.}

\item{...}{Further arguments passed to \code{\link[rstan]{stan}} (e.g.
  \code{iter}, \code{chains}, \code{refresh}, etc.)}
}
\value{
An object of class \code{"stanreg"}, which is a list containing the
  components

\describe{
  \item{coefficients}{named vector of coefficients (posterior means)}
  \item{residuals}{residuals (of type \code{'response'}). See also
  \code{\link{residuals.stanreg}}}.
  \item{fitted.values}{the fitted mean values (for glms
  the linear predictors are transformed by the invserse link function).}
  \item{linear.predictors}{the linear fit on the link scale (for linear models
  this is the same as \code{fitted.values}).}
  \item{covmat}{variance-covariance matrix for the coefficients (estimated
  from the posterior draws.)}
  \item{y}{if requested, the \code{y} vector used.}
  \item{x}{if requested, the model matrix.}
  \item{model}{if requested, the model frame.}
  \item{family}{the \code{\link[stats]{family}} object used.}
  \item{prior.weights}{any weights supplied by the user.}
  \item{df.residual}{the residual degrees of freedom}
  \item{call}{the matched call.}
  \item{formula}{the formula supplied.}
  \item{data}{the \code{data} argument.}
  \item{prior.info}{a list with information about the prior distributions
  used.}
  \item{stanfit}{the stanfit object returned by \code{\link[rstan]{stan}}}
}

The accessor functions \code{coef}, \code{fitted}, and \code{resid}
can be used with objects of class \code{"stanreg"}. There are also
\code{vcov}, \code{confint} and \code{\link{se}} methods.
}
\description{
Full Bayesian inference for generalized linear modeling with
Gaussian, Student t, or Cauchy prior distributions for the coefficients.
}
\details{
The \code{stan_glm} function is similar in syntax to
  \code{\link[stats]{glm}} but rather than performing maximum likelihood
  estimation of generalized linear models, full Bayesian estimation is
  performed via Markov Chain Monte Carlo. The Bayesian model adds independent
  Gaussian, Student t, or Cauchy priors on the coefficients of the
  generalized linear model. The \code{stan_lm} function calls \code{stan_glm}
  with \code{family = gaussian}.
}
\examples{
\dontrun{
stan_lm(mpg ~ wt, data = mtcars)
stan_glm(mpg ~ wt, data = mtcars)

ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)
fit <- stan_lm(weight ~ group)
coef(fit)

counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
stan_glm(counts ~ outcome + treatment, family = poisson())
}
}
\seealso{
\code{\link[stats]{glm}}, \code{\link[stats]{lm}},
\code{\link[rstan]{stan}}
}

