% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stanreg_methods.R
\name{stanreg_methods}
\alias{confint.stanreg}
\alias{print.stanreg}
\alias{residuals.stanreg}
\alias{se}
\alias{se.stanreg}
\alias{stanreg_methods}
\alias{vcov.stanreg}
\title{Methods}
\usage{
\method{residuals}{stanreg}(object, ...)

\method{vcov}{stanreg}(object, ...)

se(object, parm)

\method{se}{stanreg}(object, parm)

\method{confint}{stanreg}(object, parm, level = 0.95, ...)

\method{print}{stanreg}(x, ...)
}
\arguments{
\item{object}{a \code{stanreg} object.}

\item{...}{other arguments. See Details.}

\item{parm}{a character vector of parameter names.}

\item{level}{confidence level.}
}
\description{
Methods
}
\details{
For \code{residuals}, currently the only argument that can be
  specified in \code{...} is \code{type}, which is used to indicate the type
  of residuals to be  returned. \code{type} can be one of \code{'response'},
  \code{'deviance'}, or \code{'pearson'}. If omitted, the default is
  \code{'response'} for linear models and otherwise the default is
  \code{'deviance'}.
}

